+++ src/app/profile/[userId]/linked-builds/actions/get-linked-builds.ts
@@ -1,19 +1,10 @@
 'use server'
 
-import { getServerSession } from '@/app/(utils)/auth'
 import { prisma } from '@/app/(utils)/db'
 
-export default async function getLinkedBuilds() {
-  const session = await getServerSession()
-  if (!session || !session.user?.id) {
-    return {
-      status: 401,
-      body: { message: 'Unauthorized' },
-    }
-  }
-
+export default async function getLinkedBuilds({ userId }: { userId: string }) {
   const linkedBuilds = await prisma.linkedBuild.findMany({
-    where: { createdById: session.user.id },
+    where: { createdById: userId },
     include: {
       LinkedBuildItems: true,
     },